requirements.txt
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
mediapipe==0.10.11
PyYAML==6.0.1